✨ Introduce AI and Basic Settings Renderers for improved settings management
- Added **AISettingsRenderer** and **BasicSettingsRenderer** classes to modularize the rendering of AI and basic settings, enhancing code organization and maintainability.
- Implemented various settings options including API key management, prompt styles, and user interface customization, allowing users to configure their experience more effectively.
- Introduced **SettingsUtils** for shared utility functions, such as saving settings with error handling and rendering template previews, promoting code reusability.
- Updated **SettingsTab** to utilize the new renderers, streamlining the settings display process and improving user interaction.

These changes enhance the user experience by providing a more structured and intuitive settings interface, aligning with clean code principles through modular design and low coupling.
